+++ R/weathercan-pkg.R
@@ -71,6 +71,8 @@
 #' @name weathercan-package
 #' @aliases weathercan weathercan-package
 #' @importFrom dplyr "%>%"
+#' @importFrom rlang .data
+
 NULL
 
 # Dealing with CRAN Notes due to Non-standard evaluation
@@ -86,7 +88,7 @@ NULL
         "prov", "distance", "flag", "hmdx", "pressure", "rel_hum", "temp_dew",
         "visib", "wind_chill", "variable", "value", "qual", "interval", "month",
         "start", "year", "elev", "problems",
-        "V1", "V2",
+        "V1", "V2", "tz",
         "html", "skip", "data",
         "." # piping requires '.' at times
       )